fix(ci): statvfs fields are u32 on macOS — cross-platform cast to u64
On Linux, `libc::statvfs::f_bavail` / `f_frsize` are `c_ulong` which is
`u64` on our 64-bit targets. On macOS they are `u32`. The previous commit
removed `as u64` casts to satisfy `clippy::unnecessary_cast` on Linux,
but that broke the Apple Darwin builds with `cannot multiply u32 by u64`.

Restore the casts with a local `#[allow(clippy::unnecessary_cast)]` on
the single line — the cast really is unnecessary on Linux but mandatory
on macOS, and a targeted allow is cleaner than a cfg gate.

Affected lines:
- crates/amplihack-cli/src/update/install.rs:235 (free-space preflight)
- crates/amplihack-cli/src/commands/multitask/orchestrator.rs:823 (disk check)

Verified: cargo build + cargo clippy --all-targets -- -D warnings clean
on Linux. The x86_64 / aarch64 Apple Darwin CI jobs should now compile.

+++ crates/amplihack-cli/src/commands/multitask/orchestrator.rs
@@ -820,7 +820,10 @@ export AMPLIHACK_MAX_SESSIONS='{max_sessions}'
         unsafe {
             let mut stat: libc::statvfs = std::mem::zeroed();
             if libc::statvfs(path_cstr.as_ptr(), &mut stat) == 0 {
-                let free_bytes = stat.f_bavail * stat.f_frsize;
+                // macOS `statvfs` uses `u32` fields; Linux uses `u64`. Cast
+                // both so the multiplication type-checks on every target.
+                #[allow(clippy::unnecessary_cast)]
+                let free_bytes = (stat.f_bavail as u64) * (stat.f_frsize as u64);
                 let free_gb = free_bytes as f64 / (1024.0 * 1024.0 * 1024.0);
                 if free_gb < min_free_gb {
                     bail!(
